+++ src/format/mod.rs
@@ -45,7 +45,7 @@ fn sign(value: f64) -> f64 {
 ///
 /// For convenience, Dinero provides the following rounding functions: up, down, halfUp, halfDown, halfOdd, halfEven (bankers rounding), halfTowardsZero, and halfAwayFromZero.
 ///
-pub fn to_unit(d: Dinero, digits: Option<isize>, round: Option<RoundingMode>) -> f64 {
+pub fn to_unit(d: Dinero, digits: Option<u32>, round: Option<RoundingMode>) -> f64 {
     let to_unit_factor = d.currency.base.pow(d.scale.try_into().unwrap()) as f64;
 
     let digits_expo = match digits {
